Fixed the edge-case sweep's wrong results: RAND now reproduces SQL Server's generator; EXISTS never evaluates its select list; bit, float-overflow, ROUND, CEILING, datetime rendering, DATEPART, DATE_BUCKET, DATEADD, SOUNDEX, LOWER/UPPER, LIKE, FORMAT and JSON_VALUE results match real; nvarchar-to-varchar conversion best-fits on single-byte code pages; a written GROUP BY over no rows returns none; grouped-away ROLLUP expressions read NULL; nested TOP 100 PERCENT drops its ORDER BY; LEN over MAX and NTILE are bigint; and catalog views carry real's object ids.

+++ tests/SqlServerSimulator.Tests/TextSizeTests.cs
@@ -138,7 +138,7 @@ public sealed class TextSizeTests
         using var reader = sim.ExecuteReader("set textsize 10; select datalength(v), len(v), v from t");
         IsTrue(reader.Read());
         AreEqual(100L, reader.GetInt64(0));
-        AreEqual(100, reader.GetInt32(1));
+        AreEqual(100L, reader.GetInt64(1));
         AreEqual(new string('x', 10), reader.GetValue(2));
     }
 
@@ -151,7 +151,7 @@ public sealed class TextSizeTests
             select len(@v), @v
             """);
         IsTrue(reader.Read());
-        AreEqual(100, reader.GetInt32(0));
+        AreEqual(100L, reader.GetInt64(0));
         AreEqual(new string('x', 10), reader.GetValue(1));
     }
